requirements.txt
google-api-python-client==2.90.0
google-auth-oauthlib==1.0.0
openai==0.27.8
pybtex==0.24.0
faiss-cpu==1.7.4
langchain==0.0.218
pypdf==3.11.1
